Allow peristance of tokens
The ios module required you to call the `start` method to setup the manager objects.
This makes it hard to persist tokens across restarts, since you would have to start an authentication flow to use the
`refresh` method.
This PR initializes the managers lazily when needed
The podspec is updated to pin a specific HAAPI SDK version.